refactor(modules): 拆分监控插件职责模块
将轮询、状态转换、通知、成就、菜单管理、领域状态、Steam/QQ 客户端和展示逻辑迁移到独立模块，保留原有插件入口与运行状态兼容边界。

补充模块结构及领域逻辑测试，并完善模块拆分目的与影响范围说明。

+++ src/presentation/web/statistics.py
@@ -139,7 +139,7 @@ def _build_heatmap_contributions(plugin, start_key, end_key, allowed_sids):
             date_key = session.get("date", "")
             if not start_key <= date_key <= end_key:
                 continue
-            minutes = max(0, int(session.get("duration_min", 0) or 0))
+            minutes = max(0.0, float(session.get("duration_min", 0) or 0))
             if minutes <= 0:
                 continue
             player_day = contributions.setdefault(date_key, {}).setdefault(
